Move ConnectionManager to Database namespace
This puts it closer to the Connection objects it manages.
