+++ lib/Cake/Database/ConnectionManager.php
@@ -14,7 +14,7 @@
  * @since         CakePHP(tm) v 0.10.x.1402
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Model;
+namespace Cake\Database;
 
 use Cake\Core\App;
 use Cake\Core\Configure;
